+++ docs/dqx/docs/guide/data_profiling.mdx
@@ -389,7 +389,7 @@ You can save checks defined in code or generated by the profiler to a table or f
     dq_engine.save_checks(checks, config=InstallationChecksStorageConfig(assume_user=True, run_config_name="default"))
 
     # save checks in a Delta table with default run config for filtering
-    # append checks in the table
+    # append checks in the table for the default run config
     dq_engine.save_checks(checks, config=TableChecksStorageConfig(location="dq.config.checks_table", mode="append"))
 
     # save checks in a Delta table with specific run config for filtering
