+++ docs/dqx/docs/reference/engine.mdx
@@ -84,6 +84,7 @@ Supported storage configurations (implementations of `BaseChecksStorageConfig`) 
 * `TableChecksStorageConfig` can be used to save or load checks from a table, with fields:
     * `location`: table fully qualified name
     * `run_config_name`: (optional) run configuration name to load (use "default" if not provided)
+    * `mode`: (optional) write mode for saving checks (`overwrite` or `append`, default is `overwrite`). The `overwrite` mode will only replace checks for the specific run config and not all checks in the table.
 * `InstallationChecksStorageConfig` can be used to save or load checks from workspace installation, with fields:
     * `location` (optional): automatically set based on the `checks_table` or `checks_file` field from the run configuration, depending on which is provided. If both are specified, `checks_table` takes precedence
     * `run_config_name` (optional) - run configuration name to load (use "default" if not provided).
